mqtt/pyproject.toml
[tool.poetry]
name = "mqtt"
version = "0.1.0"
description = ""
authors = ["Koray <koray.oezdemir@barco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.1.0"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
